# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: API Integration
# during dev, should point to your local machine
host: localhost:5072
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /api/getAllCmsList:
    # binds a127 app logic to a route
    x-swagger-router-controller: cms_ctrl
    post:
      tags: ['Admin Module']
      description: Get All Cms Page
      summary: Get All Cms Page
      # used as the method name of the controller
      operationId: getAllCmsList
      parameters:
        - name: body
          in: body
          description: Get All cms
          schema:
           $ref: '#/definitions/getAllCmsList'
          required: true    
        - name: Authorization
          in: header
          description: Authorization
          required: true
          type: string                          
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"                                   
  /api/updateCms:
    # binds a127 app logic to a route
    x-swagger-router-controller: cms_ctrl
    post:
      tags: ['Admin Module']
      description: Update Cms Page
      summary: Update Cms Page
      # used as the method name of the controller
      operationId: updateCms
      parameters:
        - name: body
          in: body
          description: Update Cms
          schema:
           $ref: '#/definitions/updateCms'
          required: true    
        - name: Authorization
          in: header
          description: Authorization
          required: true
          type: string                          
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"                   
  /api/getCmsByID:
    # binds a127 app logic to a route
    x-swagger-router-controller: cms_ctrl
    post:
      tags: ['Admin Module']
      description: Get Cms By ID
      summary: Get Cms By ID
      # used as the method name of the controller
      operationId: getCmsByID
      parameters:
        - name: body
          in: body
          description: Get Cms By ID
          schema:
           $ref: '#/definitions/getCmsByID'
          required: true    
        - name: Authorization
          in: header
          description: Authorization
          required: true
          type: string                          
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"                     
  /api/getCmsPageByName:
    # binds a127 app logic to a route
    x-swagger-router-controller: cms_ctrl
    post:
      tags: ['Admin Module']
      description: Get Cms By name
      summary: Get Cms By name
      # used as the method name of the controller
      operationId: getCmsPageByName
      parameters:
        - name: body
          in: body
          description: Get Cms By name
          schema:
           $ref: '#/definitions/getCmsPageByName'
          required: true                 
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /api/getUserByIdAdmin/{id}:
    # binds a127 app logic to a route
    x-swagger-router-controller: admin_ctrl
    get:
      tags: ['Admin Module']
      description: Get User By Id 
      summary:  Get User By Id 
      # used as the method name of the controller
      operationId: getUserByIdAdmin      
      parameters:
        - name: id
          in: path
          description: User Id
          required: true
          type: string        
        - name: Authorization
          in: header
          description: Authorization
          required: true
          type: string                         
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"       
...
  /api/productStatusChange:
    # binds a127 app logic to a route
    x-swagger-router-controller: admin_ctrl
    post:
      tags: ['Admin Module']
      description: Product status change
      summary: Product status change
      # used as the method name of the controller
      operationId: productStatusChange
      parameters:
        - name: body
          in: body
          description: Product status change
          schema:
           $ref: '#/definitions/productStatusChange'
          required: true    
        - name: Authorization
          in: header
          description: Authorization
          required: true
          type: string                          
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /api/deleteProduct:
    # binds a127 app logic to a route
    x-swagger-router-controller: admin_ctrl
    post:
      tags: ['Admin Module']
      description: Product Delete
      summary: Product Delete
      # used as the method name of the controller
      operationId: deleteProduct
      parameters:
        - name: body
          in: body
          description: Product Delete
          schema:
           $ref: '#/definitions/deleteProduct'
          required: true    
        - name: Authorization
          in: header
          description: Authorization
          required: true
          type: string                          
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"                                     
...
definitions:
...
  getCmsByID:
    type: object
    properties:
      id:
        type: string                
  getCmsPageByName:
    type: object
    properties:
      name:
        type: string  
  getAllCmsList:
    type: object
    properties:
      page:
        type: string 
      count:
        type: string                        
  updateCms:
    type: object
    properties:
      title:
        type: string 
      description:
        type: string                   
      _id:
        type: string                
...
  productStatusChange:
    type: object
    properties:
      product_id:
        type: string 
      status:
        type: string                
  deleteProduct:
    type: object
    properties:
      product_id:
        type: string                   
...
  appResponse:
    required:
      - code
    properties:
      code:
        type: number
      status:
        type: string        
      message:
        type: string        
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string